feat: add rename css feature
This commit introduces a change in the way CSS selectors are extracted.
Previously, only the keys of the query object were used as filters. Now,
both keys and values are used, allowing for more precise CSS extraction.
The values in the query object can now be used to replace the selected
CSS, providing more flexibility and control over the CSS extraction
process.